requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.0
sqlalchemy[asyncio]==2.0.48
alembic==1.13.1
asyncpg==0.30.0
pydantic-settings==2.3.0
google-api-python-client==2.131.0
google-auth-oauthlib==1.2.0
langchain-core==0.2.27
langgraph==0.2.0
anthropic==0.28.0
google-generativeai==0.7.0
mistralai==0.4.2
tavily-python==0.3.3
apscheduler==3.10.4
cryptography==42.0.8
aiosqlite==0.20.0
